backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pandas
openpyxl
sentence-transformers
faiss-cpu
